se añadio la funcionalidad en itemMenuServiceImpl

+++ src/main/java/com/piero/backend/chat/app/menu/dto/itemmenu/ItemMenuDTORequest.java
@@ -8,6 +8,6 @@ public record ItemMenuDTORequest(
     Double precio,
     File imagen,
     String nombreImagen,
-    Integer idCategoria
+    Short idCategoria
 ) {
 }
